Working timer code
The 1-second-tick timer is put into a function called return_a_second(). This function will return 1 when a second has passed. The while loop will have a delay of HAL_Delay(1). There will be another variable called 'timer' that will increase whenever a second has passed. The plan is to use this variable to control when stuff happens.